requirements.txt
fastapi
uvicorn
cryptography
pyjwt
pytest
httpx
